# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AuditOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /engagements:
    get:
      operationId: listEngagements
      tags: [engagements]
      summary: List all engagements
      responses:
        "200":
          description: List of engagements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Engagement"
    post:
      operationId: createEngagement
      tags: [engagements]
      summary: Create a new engagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEngagementBody"
      responses:
        "201":
          description: Created engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
  /engagements/{id}:
    get:
      operationId: getEngagement
      tags: [engagements]
      summary: Get an engagement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEngagement
      tags: [engagements]
      summary: Update an engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEngagementBody"
      responses:
        "200":
          description: Updated engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /anomalies:
    get:
      operationId: listAnomalies
      tags: [anomalies]
      summary: List GL anomalies
      parameters:
        - name: engagementId
          in: query
          schema:
            type: ["integer", "null"]
        - name: riskLevel
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of anomalies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GlAnomaly"
    post:
      operationId: createAnomaly
      tags: [anomalies]
      summary: Create a GL anomaly record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnomalyBody"
      responses:
        "201":
          description: Created anomaly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlAnomaly"
  /anomalies/{id}:
    get:
      operationId: getAnomaly
      tags: [anomalies]
      summary: Get an anomaly by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anomaly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlAnomaly"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAnomaly
      tags: [anomalies]
      summary: Update anomaly status and notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnomalyBody"
      responses:
        "200":
          description: Updated anomaly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlAnomaly"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /anomalies/{id}/analyze:
    post:
      operationId: analyzeAnomaly
      tags: [ai]
      summary: AI deep-dive analysis of an anomaly
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /data-room:
    get:
      operationId: listDataRoomItems
      tags: [data-room]
      summary: List data room items
      parameters:
        - name: engagementId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of data room items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataRoomItem"
    post:
      operationId: createDataRoomItem
      tags: [data-room]
      summary: Create a data room item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDataRoomItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRoomItem"
  /data-room/{id}:
    patch:
      operationId: updateDataRoomItem
      tags: [data-room]
      summary: Update a data room item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDataRoomItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRoomItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        tin:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        financialYearEnd:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        tin:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        financialYearEnd:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
      required:
        - name
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        tin:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        financialYearEnd:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        isActive:
          type: boolean
    Engagement:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        engagementName:
          type: string
        periodStart:
          type: string
        periodEnd:
          type: string
        engagementType:
          type: string
        status:
          type: string
        partnerId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        materialityUsd:
          type: ["number", "null"]
        performanceMatUsd:
          type: ["number", "null"]
        overallRisk:
          type: string
        reportDate:
          type: ["string", "null"]
        signedOffAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        client:
          $ref: "#/components/schemas/Client"
      required:
        - id
        - clientId
        - engagementName
        - periodStart
        - periodEnd
        - engagementType
        - status
        - overallRisk
        - createdAt
        - updatedAt
        - client
    CreateEngagementBody:
      type: object
      properties:
        clientId:
          type: integer
        engagementName:
          type: string
        periodStart:
          type: string
        periodEnd:
          type: string
        engagementType:
          type: string
        overallRisk:
          type: string
        partnerId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        materialityUsd:
          type: ["number", "null"]
        reportDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - clientId
        - engagementName
        - periodStart
        - periodEnd
        - engagementType
        - overallRisk
    UpdateEngagementBody:
      type: object
      properties:
        engagementName:
          type: string
        status:
          type: string
        overallRisk:
          type: string
        materialityUsd:
          type: ["number", "null"]
        reportDate:
          type: ["string", "null"]
        signedOffAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        partnerId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
    GlAnomaly:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        anomalyType:
          type: string
        riskLevel:
          type: string
        accountCode:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        journalRef:
          type: ["string", "null"]
        transactionDate:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        aiExplanation:
          type: ["string", "null"]
        auditorStatus:
          type: string
        auditorNotes:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - anomalyType
        - riskLevel
        - auditorStatus
        - createdAt
    CreateAnomalyBody:
      type: object
      properties:
        engagementId:
          type: integer
        anomalyType:
          type: string
        riskLevel:
          type: string
        accountCode:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        journalRef:
          type: ["string", "null"]
        transactionDate:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        aiExplanation:
          type: ["string", "null"]
      required:
        - engagementId
        - anomalyType
        - riskLevel
    UpdateAnomalyBody:
      type: object
      properties:
        auditorStatus:
          type: string
        auditorNotes:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
    AiAnalysisResult:
      type: object
      properties:
        analysis:
          type: string
        updatedExplanation:
          type: string
      required:
        - analysis
        - updatedExplanation
...
    DataRoomItem:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        category:
          type: string
        itemName:
          type: string
        description:
          type: ["string", "null"]
        requestedAt:
          type: string
          format: date-time
        dueDate:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        status:
          type: string
        requestedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - category
        - itemName
        - requestedAt
        - status
        - createdAt
    CreateDataRoomItemBody:
      type: object
      properties:
        engagementId:
          type: integer
        category:
          type: string
        itemName:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        requestedBy:
          type: ["integer", "null"]
      required:
        - engagementId
        - category
        - itemName
    UpdateDataRoomItemBody:
      type: object
      properties:
        status:
          type: string
        receivedAt:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]